requirements.txt
fastapi>=0.115.11
uvicorn>=0.34.0
supabase>=2.13.0
httpx>=0.26
jinja2>=3.1.6
python-multipart>=0.0.20
pytest>=8.3.5
pytest-asyncio>=0.25.3
openai>=1.14.2
python-dotenv>=1.0.1
email-validator>=2.2.0
python-jose[cryptography]>=3.4.0